# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de gastos e informações financeiras",
        "title": "Contabius API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://vkunssec.vercel.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/vkunssec/contabius/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "${HOST}",
    "basePath": "/",
    "paths": {
        "/account": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para retornar as contas bancárias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Rota para retornar as contas bancárias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "678079f6f5080a39a8eedc1e"
                        ],
                        "description": "Ids das contas bancárias a serem retornadas",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Accounts"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para criar uma conta bancária",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Rota para criar uma conta bancária",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da conta bancária",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/account/remove/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para deletar uma conta bancária",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Rota para deletar uma conta bancária",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da conta bancária",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/account/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para atualizar uma conta bancária",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Rota para atualizar uma conta bancária",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da conta bancária",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Accounts"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID da conta bancária",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/category": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para retornar todas as categorias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Rota para retornar todas as categorias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "678079f6f5080a39a8eedc1e"
                        ],
                        "description": "Ids das categorias a serem retornadas",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetCategoriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para criar uma categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Rota para criar uma categoria",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da categoria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerError"
                        }
                    }
                }
            }
        },
        "/category/remove/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para deletar uma categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Rota para deletar uma categoria",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da categoria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AccountRequest": {
            "type": "object",
            "required": [
                "account",
                "color"
            ],
            "properties": {
                "account": {
                    "description": "Nome da conta",
                    "type": "string",
                    "example": "Conta Corrente"
                },
                "color": {
                    "description": "Cor da conta",
                    "type": "string",
                    "example": "#000000"
                }
            }
        },
        "Accounts": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "Nome da conta",
                    "type": "string",
                    "example": "Conta Corrente"
                },
                "color": {
                    "description": "Cor da conta",
                    "type": "string",
                    "example": "#000000"
                },
                "created_at": {
                    "description": "Data de criação",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID da conta",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "updated_at": {
                    "description": "Data de atualização",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "BadRequest": {
            "description": "Estrutura padrão de resposta de erro de requisição",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Erro ao processar a requisição"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "Categories": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Nome da categoria",
                    "type": "string",
                    "example": "Alimentação"
                },
                "created_at": {
                    "description": "Data de criação",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID da categoria",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "parent": {
                    "description": "ID do pai da categoria",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "updated_at": {
                    "description": "Data de atualização",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "CategoryRequest": {
            "type": "object",
            "required": [
                "category"
            ],
            "properties": {
                "category": {
                    "description": "Nome da categoria",
                    "type": "string",
                    "example": "Alimentação"
                },
                "parent": {
                    "description": "ID do pai da categoria",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                }
            }
        },
...
        "GetCategoriesResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "description": "Dados das categorias",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Categories"
                    }
                },
                "message": {
                    "description": "Mensagem de sucesso ou erro",
                    "type": "string",
                    "example": "Categorias retornadas com sucesso"
                },
                "success": {
                    "description": "Sucesso da operação",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "InternalServerError": {
            "description": "Estrutura padrão de resposta de erro de servidor interno",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "Response": {
            "description": "Estrutura padrão de resposta",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Dados da resposta"
                },
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}